fix: test chaining
reset SV axi pointers immediately
Otherwise wptr_nxt was being called down the line in another thread, and in the meanwhile SV has come out of reset and caused spurious read data to be returned over axi

Unfortunately wptr_nxt can't be reset in SV due to a zebu bug. Zebu seems to have an issue if the signal is driven by both a DPI export and an always block.