fix(zebu): axi_sw_r_8 was not working
After the change from axi_sw_r to axi_sw_r_{8,64}, the zebu started being flakey with axi_sw_r*.
Correct read data was not being returned to the part.
Tried a lot of methods, but this ugly way of removing a layer of hierarchy seemed to make it work.