feat(emu): replace modulos with counters
These are sometimes on the zebu critical path
I verified that the cycle based sync looks exactly the same in vcs as previously
Instruction based sync is a little bit different. The previous math was a little weird. Now I think it makes more sense, there'll be a sync on the cycle that instructions crosses period. Sometimes it happened the cycle before previously