Fix Verilator waveform time points
Previously Verilator waveform dumps used to come out with the time points incorrect in the dumps due to the simulation time being incremented after a call to 'eval' but before the trace 'dump' call corresponding to that eval. This used to manifest as the clock signals being "squiggly". Now they are proper square waves as intended.